Liwei

+++ WebSite/src/components/StallPage/CommentCardSection.vue
@@ -16,23 +16,23 @@
       </q-item>
 
       <q-separator/>
-      
+
       <q-card-section>
         <div class="ellipsis-2-lines" >
             {{ reviewComment }}
         </div>
       </q-card-section>
-      
+
       <q-card-section vertical class="q-pt-none" >
         <q-img :src="reviewImage"/>
       </q-card-section>
 
-      
+
 
       <q-card-section class="q-pt-none ">
         <q-btn size="sm" falt round color="primary" icon="thumb_up" @click="PostreviewLikes(reviewID)" />
         <span class="q-px-sm text-caption text-grey"> {{ reviewLikes }} </span>
-    
+
       <q-btn label="More" flat color="grey"  dense @click="dialog = true" >
 
           <q-dialog v-model="dialog" persistent>
@@ -63,14 +63,14 @@ import { ref, defineComponent } from 'vue'
 import axios from "axios";
 
 export default defineComponent({
-  
+
     setup () {
     return {
       dialog: ref(false),
       islike: ref(false),
       }
     },
-  
+
     name: "CommentCardSection",
     props:{
         reviewID: {
@@ -105,33 +105,39 @@ export default defineComponent({
             type:String,
             default:'#'
         },
-        
+
     },
      methods:{
       PostreviewLikes(ID){
         var reviewID = ID;
         var API_LINK = `http://localhost:3000/reviews/like/${reviewID}`
         console.log(this.islike)
-        if(this.islike == true){
-          axios.delete(API_LINK,NaN,{headers:{Authorization:"token 9944b09199c62b4bbdfc6ee4b"}}).then(function(response){
-            console.log(response.data)
-            this.islike = false
+                    if(!this.$store.state.login.loginStatus){
+                      this.$router.push('/login')
+                    }
+                    else{
+                      if(this.islike == true){
+                        axios.delete(API_LINK,NaN,{headers:{Authorization:"token 9944b09199c62b4bbdfc6ee4b"}}).then(function(response){
+                          console.log(response.data)
+                          this.islike = false
 
-            this.$router.go(0)
-          });
-        }
-        else if(this.islike == false){
-          axios.post(API_LINK,NaN,{headers:{Authorization:"token 9944b09199c62b4bbdfc6ee4b"}}).then(function(response){
-            console.log(response.data)
-            this.islike = true
+                          this.$router.go(0)
+                        });
+                      }
+                      else if(this.islike == false){
+                        axios.post(API_LINK,NaN,{headers:{Authorization:"token 9944b09199c62b4bbdfc6ee4b"}}).then(function(response){
+                          console.log(response.data)
+                          this.islike = true
 
-            this.$router.go(0)
-          });
-        }
+                          this.$router.go(0)
+                        });
+                      }
+                    }
+       
       }
-  
+ 
     }
-    
+
 })
 </script>
